Funcoes adicionadas Ints
randomKeyValueShort()
alloc_matrix_int()
store_key_int()
exists_key_int()
get_private_key_int()
get_runlength_int()
delete_key_int()
bulk_populate_public_keys_int()
bulk_compute_private_keys_int()
bulk_compute_runlengths_int()

+++ lib_projeto/projeto.h
@@ -19,9 +19,9 @@ struct matrixString{
 };
 
 struct matrixInts{
-    int **matrixPub;
-    int **matrixPriv;
-    int **matrixCod;
+    short **matrixPub;
+    short **matrixPriv;
+    short **matrixCod;
 };
 
 int bipolar_number(char *text);
